+++ sandboxes/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @adobe/alloy-sandbox-node
 
+## 0.0.4-beta.1
+
+### Patch Changes
+
+- Updated dependencies [[`34fab83`](https://github.com/adobe/alloy/commit/34fab83b4aa76fc7c1c023c47a1bd5a7f968082d)]:
+  - @adobe/alloy-node@0.1.0-beta.1
+
 ## 0.0.4-beta.0
 
 ### Patch Changes